Move validation functions to tests/utils.py
